refactor pallet configs
