+++ templates/parachain/runtime/src/configs/authorship.rs
@@ -1,0 +1,6 @@
+use crate::*;
+
+impl pallet_authorship::Config for Runtime {
+	type FindAuthor = pallet_session::FindAccountFromAuthorIndex<Self, Aura>;
+	type EventHandler = (CollatorSelection,);
+}
