+++ templates/parachain/runtime/src/configs/sudo.rs
@@ -1,0 +1,7 @@
+use crate::*;
+
+impl pallet_sudo::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type RuntimeCall = RuntimeCall;
+	type WeightInfo = ();
+}
